Minor modifications to reduce the number of warnings issued (particulary warnings in header files).

+++ libs/ofxCv/src/Helpers.cpp
@@ -2,9 +2,9 @@
 #include "ofxCv/Utilities.h"
 
 namespace ofxCv {
-	
+
 	using namespace cv;
-	
+
 	ofMatrix4x4 makeMatrix(Mat rotation, Mat translation) {
 		Mat rot3x3;
 		if(rotation.rows == 3 && rotation.cols == 3) {
@@ -19,11 +19,11 @@ namespace ofxCv {
 											 rm[2], rm[5], rm[8], 0.0f,
 											 tm[0], tm[1], tm[2], 1.0f);
 	}
-	
+
 	void drawMat(Mat& mat, float x, float y) {
 		drawMat(mat, x, y, mat.cols, mat.rows);
 	}
-	
+
     // experimental special case of copying into ofTexture, which acts different
     // might be able to rewrite this in terms of getDepth() but right now this
     // function loses precision with CV_32F -> CV_8U
@@ -47,7 +47,7 @@ namespace ofxCv {
 		tex.allocate(w, h, glType);
 		tex.loadData(buffer.ptr(), w, h, glType);
     }
-    
+
 	void drawMat(Mat& mat, float x, float y, float width, float height) {
         if(mat.empty()) {
             return;
@@ -56,38 +56,38 @@ namespace ofxCv {
         copy(mat, tex);
 		tex.draw(x, y, width, height);
 	}
-	
+
 	void applyMatrix(const ofMatrix4x4& matrix) {
 		glMultMatrixf((GLfloat*) matrix.getPtr());
 	}
-	
+
 	int forceOdd(int x) {
 		return (x / 2) * 2 + 1;
 	}
-	
+
 	int findFirst(const Mat& arr, unsigned char target) {
-		for(int i = 0; i < arr.rows; i++) {
+		for(unsigned int i = 0; i < arr.rows; i++) {
 			if(arr.at<unsigned char>(i) == target) {
 				return i;
 			}
 		}
 		return 0;
 	}
-	
+
 	int findLast(const Mat& arr, unsigned char target) {
-		for(int i = arr.rows - 1; i >= 0; i--) {
+		for(unsigned int i = arr.rows - 1; i >= 0; i--) {
 			if(arr.at<unsigned char>(i) == target) {
 				return i;
 			}
 		}
 		return 0;
 	}
-	
+
 	float weightedAverageAngle(const vector<Vec4i>& lines) {
 		float angleSum = 0;
 		ofVec2f start, end;
 		float weights = 0;
-		for(int i = 0; i < lines.size(); i++) {
+		for(unsigned int i = 0; i < lines.size(); i++) {
 			start.set(lines[i][0], lines[i][1]);
 			end.set(lines[i][2], lines[i][3]);
 			ofVec2f diff = end - start;
@@ -99,19 +99,19 @@ namespace ofxCv {
 		}
 		return angleSum / weights;
 	}
-	
+
 	vector<cv::Point2f> getConvexPolygon(const vector<cv::Point2f>& convexHull, int targetPoints) {
 		vector<cv::Point2f> result = convexHull;
-		
+
 		static const unsigned int maxIterations = 16;
 		static const double infinity = numeric_limits<double>::infinity();
 		double minEpsilon = 0;
 		double maxEpsilon = infinity;
 		double curEpsilon = 16; // good initial guess
-		
+
 		// unbounded binary search to simplify the convex hull until it's targetPoints
-		if(result.size() > targetPoints) {
-			for(int i = 0; i < maxIterations; i++) {
+		if(result.size() > (unsigned int) targetPoints) {
+			for(unsigned int i = 0; i < maxIterations; i++) {
 				approxPolyDP(Mat(convexHull), result, curEpsilon, true);
 				if(result.size() == targetPoints) {
 					break;
@@ -130,18 +130,18 @@ namespace ofxCv {
 				}
 			}
 		}
-		
+
 		return result;
 	}
-	
+
 	void drawHighlightString(string text, ofPoint position, ofColor background, ofColor foreground) {
 		drawHighlightString(text, position.x, position.y, background, foreground);
 	}
-	
+
 	void drawHighlightString(string text, int x, int y, ofColor background, ofColor foreground) {
 		vector<string> lines = ofSplitString(text, "\n");
 		int textLength = 0;
-		for(int i = 0; i < lines.size(); i++) {
+		for(unsigned int i = 0; i < lines.size(); i++) {
 			// tabs are not rendered
 			int tabs = count(lines[i].begin(), lines[i].end(), '\t');
 			int curLength = lines[i].length() - tabs;
@@ -153,13 +153,13 @@ namespace ofxCv {
 				textLength = curLength;
 			}
 		}
-		
+
 		int padding = 4;
 		int fontSize = 8;
 		float leading = 1.7;
 		int height = lines.size() * fontSize * leading - 1;
 		int width = textLength * fontSize;
-		
+
 #ifdef TARGET_OPENGLES
 		// This needs to be refactored to support OpenGLES
 		// Else it will work correctly
@@ -178,7 +178,7 @@ namespace ofxCv {
 		ofPopMatrix();
 		ofPopStyle();
 		glPopAttrib();
-#endif         
-         
+#endif
+
 	}
 }
